Run full multi-prompt cartography sweep on Qwen 7B
Optimized perturbation engine (compute baseline once, reuse across all 784 heads).
Fixed MLX mask dtype (float16, not bfloat16) for updated mlx-lm.
Ran 5-prompt sweep: greeting, safety, reasoning, creative, identity.
Key finding: head.27.2 is the universal output bottleneck, 46% of heads are inert.

+++ src/heinrich/probe/mlx_provider.py
@@ -71,9 +71,9 @@ class MLXProvider:
                 h = embed_fn(input_ids)
                 hidden_states.append(np.array(h.astype(mx.float32)[0, -1, :]))
 
-                # Causal mask must be bfloat16 to match model dtype
+                # Causal mask must match model's attention dtype (float16 for quantized models)
                 T = h.shape[1]
-                mask = mx.triu(mx.full((T, T), float('-inf'), dtype=mx.bfloat16), k=1) if T > 1 else None
+                mask = mx.triu(mx.full((T, T), float('-inf'), dtype=mx.float16), k=1) if T > 1 else None
 
                 for layer in layers:
                     h = layer(h, mask=mask, cache=None)
